requirements.txt
backcall==0.2.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
decorator==4.4.2
distro==1.5.0
gpiozero==1.6.2
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.3
MarkupSafe==2.0.1
meson==0.61.1
Mopidy==3.1.1
Mopidy-MPD==3.2.0
Mopidy-MusicBox-Webclient==3.1.0
Mopidy-YouTube==3.5
numpy==1.19.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
ply==3.11
prompt-toolkit==3.0.14
pyalsaaudio==0.8.4
pyFluidSynth==1.3.0
Pygments==2.7.1
PyGObject==3.38.0
Pykka==2.0.3
python-apt==2.2.1
python-mpd2==3.0.4
python-musicpd==0.7.0
PyYAML==6.0
pyzbar==0.1.8
requests==2.25.1
RPi.GPIO==0.7.0
six==1.16.0
soupsieve==2.3.1
spidev==3.5
ssh-import-id==5.10
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.1.9
youtube-dl==2021.12.17
ytmusicapi==0.20.0
